requirements.txt
google-genai>=1.0.0
uvicorn>=0.30.0
fastapi>=0.111.0
websockets>=12.0
python-dotenv>=1.0.0
